requirements.txt
# Django and core
Django>=5.0,<6.0
psycopg2-binary>=2.9.9
python-decouple>=3.8

# Authentication and security
argon2-cffi>=23.1.0

# HTMX integration
django-htmx>=1.17.0

# Filtering
django-filter>=24.0

# Image processing (for attachments)
Pillow>=10.0.0

# Background tasks
django-q2>=1.6.0

# Development tools
django-debug-toolbar>=4.2.0

# Testing
pytest>=8.0.0
pytest-django>=4.7.0
factory-boy>=3.3.0
